+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/Pages/Project/Tasks/ImpactAssessment/EditImpactAssessmentTask.cshtml.cs
@@ -7,7 +7,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
 using System;
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
+using Dfe.ManageFreeSchoolProjects.Models;
 
 namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.ImpactAssessment
 {
@@ -20,15 +23,22 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.ImpactAssessment
 
         [BindProperty(SupportsGet = true, Name = "projectId")]
         public string ProjectId { get; set; }
-        
+
         [BindProperty(Name = "impact-assessment-done")]
-        
-        public bool? ImpactAssessmentDone{ get; set; }
-        
+        public bool? ImpactAssessmentDone { get; set; }
+
         [BindProperty(Name = "saved-to-workplaces")]
-        
         public bool? SavedToWorkplaces { get; set; }
-        
+
+        [BindProperty(Name = "sent-section9-letter-to-local-authority")]
+        [DisplayName("Sent section 9 letter to local authority")]
+        public bool SentSection9LetterToLocalAuthority { get; set; }
+
+        [BindProperty(Name = "date-sent", BinderType = typeof(DateInputModelBinder))]
+        [DisplayName("Date sent")]
+        public DateTime? Section9LetterDateSent { get; set; }
+
+
         public string SchoolName { get; set; }
 
         public EditImpactAssessmentTaskModel(IGetProjectByTaskService getProjectService,
@@ -55,15 +65,33 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.ImpactAssessment
             var project = await _getProjectService.Execute(ProjectId, TaskName.ImpactAssessment);
             SchoolName = project.SchoolName;
 
+            if (SentSection9LetterToLocalAuthority)
+            {
+                if (Section9LetterDateSent == null)
+                {
+                    ModelState.AddModelError("date-sent", "Enter a date sent");
+                }
+            }
+            else
+            {
+                Section9LetterDateSent = null;
+            }
+            
+            if (!ModelState.IsValid)
+            {
+                _errorService.AddErrors(ModelState.Keys, ModelState);
+                return Page();
+            }
+            
             try
             {
-                var request = new UpdateProjectByTaskRequest()
+                var request = new UpdateProjectByTaskRequest
                 {
-                    ImpactAssessment = new ImpactAssessmentTask()
+                    ImpactAssessment = new ImpactAssessmentTask
                     {
                         ImpactAssessment = ImpactAssessmentDone,
-                        SavedToWorkplaces = SavedToWorkplaces
-                        
+                        SavedToWorkplaces = SavedToWorkplaces,
+                        Section9LetterDateSent = Section9LetterDateSent,
                     }
                 };
 
@@ -83,9 +111,10 @@ namespace Dfe.ManageFreeSchoolProjects.Pages.Project.Tasks.ImpactAssessment
 
             ImpactAssessmentDone = project.ImpactAssessment.ImpactAssessment;
             SavedToWorkplaces = project.ImpactAssessment.SavedToWorkplaces;
-          
-           
+            Section9LetterDateSent = project.ImpactAssessment.Section9LetterDateSent;
             SchoolName = project.SchoolName;
+
+            SentSection9LetterToLocalAuthority = Section9LetterDateSent != null; 
         }
     }
 }